feat(marketing): social proof signup count
